#1486 - moved LinksEditView.tsx to ProjectForm folder
